Make inferior and thread Python objects inherit from PyObject
This will remove the need to have casts when trying to pass an
inferior_object/thread_object pointer to something that accepts a
PyObject pointer.

This could be done for all of our custom Python types, not just inferior
and thread.  However, I want to see if everything is going fine before
taking the time to convert all of them.

gdb/ChangeLog:

	* python/py-inferior.c (struct threadlist_entry,
	inferior_object, ): Move to python-internal.h.
	(inferior_to_inferior_object, infpy_threads): Remove casts.
	* python/py-infthread.c (gdbpy_selected_thread): Likewise.
	* python/py-threadevent.c (get_event_thread): Likewise.
	* python/python-internal.h (struct thread_object): Remove
	typedef, inherit from PyObject.
	(struct threadlist_entry): Move here.
	(struct inferior_object): Move here, remove typedef, inherit from
	PyObject.