Make gdb_continue_to_breakpoint accept an expected breakpoint number
The gdb_continue_to_breakpoint procedure currently accepts a location
pattern that is checked against the function name/address/filename part
of the breakpoint hit message.  This patch adds a similar parameter, but
for the breakpoint number.  It can be used to verify that the execution
has stopped at the breakpoint we are expected to hit.

gdb/testsuite/ChangeLog:

	* lib/gdb.exp (gdb_continue_to_breakpoint): Add breakpoint_num
	parameter, use it in the regex.