Add new test for thread-specific breakpoints with conditions
While developing the gdbserver-based thread-specific breakpoints, I had
a few bugs in my implementation, which lead me to write these tests.  I
think it would be useful to have them in the testsuite, since it was
easy to get something wrong with the generation of the condition.

gdb/testsuite/ChangeLog:

	* gdb.threads/thread-specific-plus-condition.exp: New file.
	* gdb.threads/thread-specific-plus-condition.c: New file.