Call _initialize_printcmd and _initialize_cli_cmds early
The following commit:

  b593ecca856860a8b38deb808493bba4beef3aee
  Makefiles: Flatten and sort file lists

had the unintended effect of changing the order in which _initialize
functions are called.  A consequence is that the "tty" command (an alias
to "set inferior-tty") does not work anymore.  More precisely, it's
because _initialize_printcmd is now called after _initialize_infcmd.

Before this problem was introduced, these were the steps leading to
"tty" being successfully registered:

1. The "set" command is registered as a prefix command in printcmd.c,
which adds it to the global command list (cmdlist).  setlist is the list
of its subcommands.

2. The "set inferior-tty" command is added as a child of "set" (i.e.
it's inserted into setlist) in infcmd.c.

3. The "tty" alias is created for "set inferior-tty" in infcmd.c.  This
looks up "set" in cmdlist, then "inferior-tty" in the subcommands of
"set".  It's found and everyone is happy.

With the _initialize functions being called in the reverse order, these
steps are executed in the order 2-3-1.  When we try to create the alias,
the "set" command has not been created yet and is therefore not part of
cmdlist.  We can't find the "set inferior-tty" command, so the alias is
not created

The way I'd like to fix this for the 8.0 branch is to simply change the
order in which the _initialize functions are called.  This is not
pretty, but it avoids doing some big code changes.

To see if there would be other problems like this one, I added an assert
in add_alias_cmd while testing locally and on the buildbot, to make sure
that the target command was always found.  I also looked manually in the
-nat.c files.   This is the only instance of this problem I could find.

gdb/ChangeLog:

	* Makefile.in (init.c): Call initializer of printcmd and
	cli_cmds early.

gdb/testsuite/ChangeLog:

	* gdb.base/set-inferior-tty.exp: Test using both
	"set inferior-tty" and "tty" commands.
	(test_set_inferior_tty): Add command parameter.