Read stop_pc after the gdbarch update when exec'ing
When an inferior execs and changes architecture (e.g. 64 bits to 32
bits), the gdbarch associated to the inferior is updated by the
follow_exec call in handle_inferior_event_1.  However, we read the
stop_pc just before that.  At this point, the gdbarch of the inferior
may not match the architecture of the process.  As a result, we try to
interpret the registers sent by the remote using the wrong architecture.

This problem is exposed by the gdb.multi/multi-arch-exec.exp, which this
patch fixes.

This bug was hiding another problem, and as such introduces some
failures in gdb.base/foll-exec-mode.exp.  The following patch takes care
of that.

gdb/ChangeLog:

	* infrun.c (handle_inferior_event_1): When exec'ing, read
	stop_pc after follow_exec.