Add thread after updating gdbarch when exec'ing
When a process does an "exec" with "follow-exec-mode new", a new
inferior is added by follow_exec.  The gdbarch of that new inferior is
at first set to the default architecture of the gdb build, which may not
be the right one.  It will be updated later by the call to
target_find_description.  Before that point, it is not safe to read the
inferior's registers, because we may not interpret them correctly.  This
is exposed by a failure in gdb.base/foll-exec-mode.exp, with:

  Remote 'g' packet reply is too long (expected 312 bytes, got 816 bytes)

The call to "add_thread" done just after adding the inferior is
problematic, because it ends up reading the registers (actually, it's
switch_to_thread which tries to update stop_pc).  The registers returned
by gdbserver are the x86-64 ones, while we try to interpret them using
the "i386" gdbarch.

This issue was hidden by the bug fixed by the previous patch.  What I
think happened before that patch is that we were updating stop_pc before
switching to the new inferior, filling the regcache associated to the
thread (this worked fine as long as the architectures of the previous
and new process images were the same).  The call to switch_to_thread
worked, because the register read hit the regcache.  Now it triggers a
register read, while the gdbarch is not set correctly, leading to the
"reply is too long" error *.

Postponing the call to add_thread to until the target
description/gdbarch has been updated seems to fix the issue.

gdb/ChangeLog:

	* infrun.c (follow_exec): Call add_thread after
	target_find_description.