Show optimized out local variables in "info locals"
Currently, optimized out variables are not shown when doing "info
locals".  Some users found that confusing, thinking GDB forgot to print
their variable.  This patch adds them to the "info locals" output.  I
changed gdb.dwarf2/var-access.exp to test for that.

Doing so reveals what I think is a bug in GDB, see:

http://lists.dwarfstd.org/pipermail/dwarf-discuss-dwarfstd.org/2017-September/004394.html

This patch marks the tests in inline-locals.exp that start failing as
KFAIL.  I haven't created a bug in bugzilla yet, but if we agree it's
indeed a bug,  I'll create one and update the setup_kfail lines with the
actual bug number before pushing this patch.

gdb/ChangeLog: