C++-ify parse_format_string
This patch C++ifies the parse_format_string function and its return
value, allowing to get rid of free_format_pieces_cleanup.  Currently,
format_piece::string points into a big buffer shared by all format
pieces resulting of the format string parse.  To make the code and
allocation scheme simpler to understand, this patch makes each piece own
its string with an std::string.

I've tried to measure the impact of this change on the time taken by
parse_format_string in an optimized build.  I ran

  $ perf record -g -- ./gdb -nx -batch -x test.gdb

with test.gdb:

  set $a = 0
  while $a < 20000
    printf "%daaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa<16 times>\n", 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16
    set $a = $a + 1
  end

The purpose of the long string of a's is to have some format pieces for
which the substring doesn't fit in the std::string local buffer.  The
command "perf report" shows this for current master:

  1.81%     1.81%  gdb  gdb  [.] parse_format_string

and this with the patch:

  1.69%     1.69%  gdb  gdb  [.] parse_format_string

So it doesn't seem like it makes much of a difference.  If you have tips
on how to use perf to get more precise measurements, I would be glad to
hear them.

gdb/ChangeLog:

	* common/format.h (struct format_piece): Add constructor.
	<string>: Change type to std::string.
	(parse_format_string): Return an std::vector of format_piece.
	(free_format_pieces): Remove.
	(free_format_pieces_cleanup): Remove.
	* common/format.c (parse_format_string): Return an std::vector of
	format_piece.  Don't allocate a big buffer, adjust to those
	changes.
	(free_format_pieces): Remove.
	(free_format_pieces_cleanup): Remove.
	* ax-gdb.c (gen_printf): Remove format_piece * parameter.
	(maint_agent_printf_command): Adjust to parse_format_string
	changes.
	* ax-gdb.h (struct format_piece): Don't forward-declare.
	(gen_printf): Remove format_piece * parameter.
	* breakpoint.c (parse_cmd_to_aexpr): Adjust to
	parse_format_string chagnes.
	* printcmd.c (ui_printf): Likewise.

gdb/gdbserver/ChangeLog:

	* ax.c (ax_printf): Adjust to parse_format_string changes.