Change {dwarf2_cu,compunit_symtab}::call_site_htab to be std::unordered_map
This patch changes the call_site_htab maps to be std::unordered_map.
The main motivation for this is that I am not too confident in my
knowledge of htab to implement the following patch with call_site_htab
as an htab.  I thought changing it to a standard container would help
with that, and would add some type safety as well.

gdb/ChangeLog:

	* symtab.h (struct compunit_symtab) <call_site_htab>: Change
	type to std::unordered_map.
	* dwarf2read.c (struct dwarf2_cu) <call_site_htab>: Change type
	to std::unordered_map.
	(process_full_comp_unit): Use std::move.
	(read_call_site_scope): Update to std::unordered_map change.
	(free_heap_comp_unit): Call delete to free cu.
	* block.c (call_site_for_pc): Update to std::unordered_map
	change.