C++ify dwarf2_cu
As a preparation for the following patches, this patch makes dwarf2_cu a
C++ object.  It initializes the fields directly (except the bitfields,
which I don't think can be done this way, so are done through the
initialization list). init_one_comp_unit becomes the constructor, and
all instanciations are updated.

gdb/ChangeLog:

	* dwarf2read.c (struct dwarf2_cu): Add constructor, initialize
	fields.
	(init_tu_and_read_dwo_dies): Use new.
	(init_cutu_and_read_dies): Use new.
	(init_cutu_and_read_dies_no_follow): Use constructor.
	(init_one_comp_unit): Rename to...
	(dwarf2_cu::dwarf2_cu): ... this, and update.