Reset objfile's section_offsets when inferior exits
I see a failure in gdb.ada/exec_changed.exp, when the test case is
compiled with a toolchain that uses position-independent code by default
(for example gcc-ada 7.2.1-2 on Arch Linux).  I think it points out that
we

The test case compiles two programs (first and second) and does this:

1. Copy "first" to "common"
2. Load "common"
3. Start the inferior, verify it lands on first's main procedure (also
   called first).
4. Copy "second" to "common"
5. Start the inferior again, verify it lands on second's main procedure
   (also called second).

The failure happens in #5, the second "start" stops on the "main"
procedure.  If I understand correctly, this procedure is boilerplate
code that the toolchain puts, not user code.