Fix printing flags field
As reported in PR gdb/22938, the field value in flags is not correct.
It is wrong in bit shift in val_print_type_code_flags, but the field
value is still correct if field_len is 1.

This patch also adds a self test to default_print_one_register_info, to
check that eflags on x86 and cpsr on aarch64 are correctly printed.

Regression tested on x86_64-linux and aarch64-linux.

gdb:

2018-03-09  Yao Qi  <yao.qi@linaro.org>

	PR gdb/22938
	* infcmd.c (test_default_print_one_register_info): New function.
	(_initialize_infcmd): Register the unit test.
	* valprint.c (val_print_type_code_flags): Fix field value.